% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_address.R
\name{authors_address}
\alias{authors_address}
\title{Parses out address information and splits it into its respective parts.
This is an internal function used by \code{authors_clean}}
\usage{
authors_address(addresses, ID)
}
\arguments{
\item{addresses}{the addresses}

\item{ID}{the authorID}
}
\description{
\code{authors_address} This function takes the output from 
\code{references_read} and pulls out address information. Splitting it into
university, department, city, state, etc.
}
