% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BITR.R
\docType{data}
\name{BITR}
\alias{BITR}
\title{Data from the journal BioTropica (pulled from Web of Knowledge)}
\format{A data frame with 10 rows and 32 variables:
\describe{
  \item{filename}{the original filename the text was created from}
  \item{refID}{the unique identifier given to each reference article by references_read()}
The remaining column codes are described on the Web of Knowledge website:
\url{https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html}
}}
\usage{
BITR
}
\description{
A dataset containing 10 articles taken from the BioTropica journal. 
This dataset represents the typical formatted output from \code{references_read()}
in the refsplitr package. It serves as a testbed for commonly miscategorized names
}
\keyword{datasets}
