% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_parse.R
\name{.authors_parse}
\alias{.authors_parse}
\title{Parses out each individual authors information from the reference information created by references_read. This is the first step in parsing out author information.}
\usage{
.authors_parse(references)
}
\arguments{
\item{references}{output from `references_read()`}
}
\description{
\code{authors_clean} This function takes the output from `references_read()` and cleans the author information.
}
\examples{
authors_parse()

}
