% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_names.R
\name{.split_names}
\alias{.split_names}
\title{This is an internal function to split names into their respective parts. It can not be called directly by the user}
\usage{
.split_names(x)
}
\arguments{
\item{x}{simply a vector containing a character of a name in the form 'last, first, middle'. Can parse out names with no middle names.

Currently it does not reliably distinguish between a second last names and middle names. In most cases it will call the first last name an additional middle name. Also jrs, srs, etc are considered middle names for the sake of ease.}
}
\description{
\code{.split_names} This is an internal function to split names into their respective parts. It can not be called directly by the user
}
