% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot_coauthor.R
\name{net_plot_coauthor}
\alias{net_plot_coauthor}
\title{Creates a network diagram of coauthors' countries linked by reference}
\usage{
net_plot_coauthor(data)
}
\arguments{
\item{addresses}{the `address` element from the list outputted from the `georef_authors()`` function, containing geocoded address latitude and longitude locations.}

\item{authors_references}{output from the read_authors() function, which links author addresses together via AU_ID.}
}
\description{
\code{net_plot_coauthor} This function takes an addresses data.frame, links it to an authors_references dataset and plots a network diagram generated for co-authorship.
}
