% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_plot_coauthor_address.R
\name{net_plot_coauthor_address}
\alias{net_plot_coauthor_address}
\title{Creates a network diagram of coauthors' addresses linked by reference, and with nodes arranged geographically}
\usage{
net_plot_coauthor_address(data, line_resolution = 10,
  mapRegion = "world")
}
\arguments{
\item{mapRegion}{what portion of the world map to show. possible values include ["world","North America","South America","Australia","Africa","Antarctica","Eurasia"]}

\item{addresses}{the `address` element from the list outputted from the `georef_authors()`` function, containing geocoded address latitude and longitude locations.}
}
\description{
\code{net_plot_coauthor_address} This function takes an addresses data.frame, links it to an authors__references dataset and plots a network diagram generated for individual points of co-authorship.
}
