% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_country.R
\name{plot_addresses_country}
\alias{plot_addresses_country}
\title{Plot addresses, the number of which are summed by country_name}
\usage{
plot_addresses_country(data, mapRegion = "world")
}
\arguments{
\item{data}{address element from the output from the \code{authors_georef()}
function, containing geocoded address latitude and longitude locations.}

\item{mapRegion}{what portion of the world map to show. possible values
include \code{"world"}, \code{"North America"}, \code{"South America"}, \code{"Australia"},
\code{"Africa"}, \code{"Antarctica"}, and \code{"Eurasia"}}
}
\description{
This function plots an addresses data.frame
object by country name.
}
\examples{

## Using the output of authors_georef (e.g., BITR_geocode)
data(BITR_geocode)
## Plots the whole world
plot_addresses_country(BITR_geocode)

## Just select North America
plot_addresses_country(BITR_geocode, mapRegion = 'North America')

}
