% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_points.R
\name{plot_addresses_points}
\alias{plot_addresses_points}
\title{Plot address point locations on world map}
\usage{
plot_addresses_points(data, mapRegion = "world")
}
\arguments{
\item{data}{the `address` element from the list outputted from the `authors_georef()`` function, containing geocoded address latitude and longitude locations.}

\item{mapRegion}{what portion of the world map to show. possible values include ["world","North America","South America","Australia","Africa","Antarctica","Eurasia"]}
}
\description{
\code{plot_addresses_points} This function plots an addresses data.frame object by point overlaid on the countries of the world.
}
\examples{
plot_addresses_points()
}
