% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_coauthor.R
\name{plot_net_coauthor}
\alias{plot_net_coauthor}
\title{Creates a network diagram of coauthors' countries linked by reference
This function takes an addresses data.frame, 
links it to an authors_references dataset and plots a network diagram 
generated for co-authorship.}
\usage{
plot_net_coauthor(data)
}
\arguments{
\item{data}{the `address` element from the list outputted from 
the `authors_georef()`` function, containing geocoded address 
latitude and longitude locations.}
}
\description{
Creates a network diagram of coauthors' countries linked by reference
This function takes an addresses data.frame, 
links it to an authors_references dataset and plots a network diagram 
generated for co-authorship.
}
\examples{
## Using the output of authors_georef (e.g., BITR_geocode)
data(BITR_geocode)
plot_net_coauthor(BITR_geocode)
}
