% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_country.R
\name{plot_net_country}
\alias{plot_net_country}
\title{Creates a network diagram of coauthors' countries linked by reference, 
#and with nodes arranged geographically}
\usage{
plot_net_country(
  data,
  lineResolution = 10,
  mapRegion = "world",
  lineAlpha = 0.5
)
}
\arguments{
\item{data}{the `address` element from the list outputted from 
the `authors_georef()`` function, containing geocoded address 
latitude and longitude locations.}

\item{lineResolution}{the resolution of the lines drawn, higher numbers will make smoother curves
default is 10.}

\item{mapRegion}{what portion of the world map to show. possible 
values include ["world","North America","South America","Australia","Africa","Antarctica","Eurasia"]}

\item{lineAlpha}{transparency of the lines, fed into ggplots alpha value. Number between 0 - 1.}
}
\description{
This function takes an addresses data.frame, 
#links it to an authors_references dataset and plots a network diagram 
generated for countries of co-authorship.
}
\examples{
## Using the output of authors_georef (e.g., BITR_geocode)
data(BITR_geocode)
## Plots the whole world
output <- plot_net_country(BITR_geocode)

## Mapping only North America
output <- plot_net_country(BITR_geocode, mapRegion = 'North America')

## Change the transparency of lines by modifying the lineAlpha parameter
output <- plot_net_country(BITR_geocode, lineAlpha = 0.2)
                 
## Change the curvature of lines by modifying the lineResolution paramater
output <- plot_net_country(BITR_geocode, lineResolution = 30 )
                 
## With all arguments: 
output <- plot_net_country(BITR_geocode, mapRegion = 'North America', lineAlpha = 0.2,
                 lineResolution = 30)



}
