% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsplitr-package.R
\docType{package}
\name{refsplitr-package}
\alias{refsplitr}
\alias{refsplitr-package}
\title{refsplitr: author name disambiguation, author georeferencing, and mapping of coauthorship networks with 'Web of Science' data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to parse and organize reference records downloaded from the 'Web of Science' citation database into an R-friendly format, disambiguate the names of authors, geocode their locations, and generate/visualize coauthorship networks. This package has been peer-reviewed by rOpenSci (v. 1.0).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/refsplitr}
  \item \url{https://docs.ropensci.org/refsplitr/}
  \item Report bugs at \url{https://github.com/ropensci/refsplitr/issues}
}

}
\author{
\strong{Maintainer}: Emilio Bruna \email{embruna@ufl.edu}

Authors:
\itemize{
  \item Auriel M.V. Fournier \email{aurielfournier@gmail.com}
  \item Matthew E. Boone
  \item Forrest R. Stevens
}

Other contributors:
\itemize{
  \item Bianca Kramer (Kramer reviewed the package (v 1.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/256>) [reviewer]
  \item Najko Jahn (Jahn reviewed the package (v1.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/256>) [reviewer]
}

}
\keyword{internal}
