% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_address.R
\name{authors_address}
\alias{authors_address}
\title{Parses out address information and splits it into its respective parts.}
\usage{
authors_address(addresses, ID)
}
\arguments{
\item{addresses}{the addresses}

\item{ID}{the authorID}
}
\description{
\code{authors_address} This function takes the output from `references_read()` and cleans the author information.
}
