% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_clean.R
\name{authors_clean}
\alias{authors_clean}
\title{Extracts the lat and long for each address}
\usage{
authors_clean(references, sim_score = 0.88, filename_root = "./",
  write_out_data = FALSE)
}
\arguments{
\item{references}{output from `references_read()`}

\item{sim_score}{default is 0.88}

\item{filename_root}{the filename root, can include relative or absolute}

\item{write_out_data}{TRUE or FALSE, do you want a .csv file written out?}
}
\description{
\code{authors_clean} This function takes the output from `references_read()` and cleans the author information.
}
