% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_clean.R
\name{authors_clean}
\alias{authors_clean}
\title{Extracts the lat and long for each address}
\usage{
authors_clean(references, sim_score = 0.88, filename_root = "./",
  write_out_data = FALSE)
}
\arguments{
\item{references}{output from `references_read()`}

\item{sim_score}{default is 0.88}

\item{filename_root}{the filename root, can include relative or absolute}

\item{write_out_data}{TRUE or FALSE, do you want a .csv file written out?}
}
\description{
\code{authors_clean} This function takes the output from `references_read()` and cleans the author information.
}
\examples{
  references <-data.frame(filename=NA, AB=NA , AF=c('Smith, Jon J.','Thompson, Bob B.','Smith,J'), AU=c('Smith, Jon J.','Thompson, Bob','Smith, J'), BP=NA , C1=c("Univ Florida, Gainesville, FL USA","University of Texas, Austin, TX, USA",NA),CR=NA,DE=NA, DI=NA, EM=c("j.smithufl.edu",NA,'jsmithusgs.gov'), EN=NA, FN=NA, FU=NA, PD=NA, PG=NA, PT=NA, PU=NA, PY=NA, RI=NA, OI=NA,PM=NA, RP=c("Univ Florida, Gainesville, FL USA","University of Texas, Austin, TX, USA",NA), SC=NA, SN=NA, SO=NA, TC=NA, TI=NA, UT=NA, VL=NA, WC=NA,Z9=NA,refID=c(1,2,3) ,stringsAsFactors=F)
df[]<-lapply(df, as.character)
authors_clean(references=references)
}
