% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_georef.R
\name{authors_georef}
\alias{authors_georef}
\title{Extracts the lat and long for each address from authors_clean}
\usage{
authors_georef(data, address_column = "address")
}
\arguments{
\item{data}{dataframe from `authors_refine()`}

\item{address_column}{name of column in quotes where the addresses are}
}
\description{
\code{authors_georef} This function takes the final author list from 
refine_authors, and calculates the lat long of the addresses. 
It does this by feeding the addresses into data science toolkit. 
In order to maximize effectiveness and mitigate errors in parsing addresses
We run this multiple times creating addresses in different ways
in hopes that the data science toolkit can recognize an address
1st. University, city, zipcode, country
2nd. City, zipcode, country
3rd. city, country
4th. University, country
}
\details{
The output is a list with three data.frames
\code{addresses} is a data frame with all information from 
refine_authors plus new location columns and calculated lat longs.
\code{missing addresses} is a data frame with all addresses could
not be geocoded
\code{addresses} is a data frame like \code{addresses} except
the missing addresses are gone.
}
\examples{

\dontrun{
BITR_georef_df <- authors_georef(BITR_refined, address_column='address')
}
}
