% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_match.R
\name{authors_match}
\alias{authors_match}
\title{Matches authors together by considering common last names, university affiliations, and emails.}
\usage{
authors_match(data, sim_score)
}
\arguments{
\item{references}{output from `references_read()`}
}
\description{
\code{authors_parse} This function requires a data.frame with 8 required columns: ID, full names, address, university, country, RI, OI, and email. It uses this information to match up common names and identify groups of people
}
\examples{
authors_match()

}
