% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_country.R
\name{plot_addresses_country}
\alias{plot_addresses_country}
\title{Plot addresses, the number of which are summed by country_name}
\usage{
plot_addresses_country(data, mapRegion = "world")
}
\arguments{
\item{data}{address element from the output from the `authors_georef()`` 
function, containing geocoded address latitude and longitude locations.}

\item{mapRegion}{what portion of the world map to show. possible values 
include ["world","North America","South America","Australia","Africa","Antarctica","Eurasia"]}
}
\description{
\code{plot_addresses_country} This function plots an addresses data.frame 
object by country name.
}
\examples{

data(BITR_geocode)

## Plots the whole world
plot_addresses_country(BITR_geocode)

## Additioinal mapRegion names can be querried using:
rworldmap::getMap()$NAME

## Just select North America
plot_addresses_country(BITR_geocode, mapRegion = 'North America')

}
