% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addresses_points.R
\name{plot_addresses_points}
\alias{plot_addresses_points}
\title{Plot address point locations on world map}
\usage{
plot_addresses_points(data, mapCountry = NULL)
}
\arguments{
\item{data}{the \code{address} element from the list output by
the `authors_georef()`` function, containing geocoded address
latitude and longitude locations.}

\item{mapCountry}{What country to map. Possible values
include \link{"USA", "Brazil", "Australia", "UK"}
use \code{data(countries) to see possible names}. No value defaults
to the world map.}
}
\description{
This function plots an addresses data.frame
object by point overlaid on the countries of the world.
}
\examples{
## Using the output of authors_georef (e.g., BITR_geocode)
data(BITR_geocode)
## Plots the whole world
plot_addresses_points(BITR_geocode)

## mapCountry names can be querried using:
data(countries)

## Plot only Brazil
plot_addresses_points(BITR_geocode, mapCountry = 'Brazil')

}
